+++ resolvers/levelResolver.go
@@ -9,10 +9,10 @@ type levelResolver struct {
 	level models.Level
 }
 
-func (l *levelResolver) Value(_ context.Context) int32 {
-	return l.level.Value
+func (l *levelResolver) Value(_ context.Context) float64 {
+	return float64(l.level.Value)
 }
 
-func (l *levelResolver) Exp(_ context.Context) int32 {
-	return l.level.Exp
+func (l *levelResolver) Exp(_ context.Context) float64 {
+	return float64(l.level.Exp)
 }
